requirements.txt
# Anthropic SDK for Claude API access
anthropic>=0.40.0

# HTTP client for Supabase API calls
aiohttp>=3.9.0

# Environment variable management
python-dotenv>=1.0.0
